Move function to the detail namespace
